+++ AirportBoards_ManualEdition/sketch.js
@@ -5,7 +5,7 @@
 // Extra for Experts:
 // 
 
-let widths = [425,200,175,125,125,250];
+let widths = [475,330,175,125,125,250];
 let airlineCodes = new Map ();
 let rows = 15;
 let cols = 6;
@@ -23,21 +23,6 @@ let cyxe, cyqr, cyyz, cyyc, cyvr, cyul;
 let klax, kjfk, kord, egll, lfpg, omdb, yssy;
 
 
-// Ignore the following code, this is to be used for pulling live data from FlightAware.
-// httpGet("http://flightxml.flightaware.com/json/FlightXML3/METHODNAME");
-// httpDo("http://flightxml.flightaware.com/json/FlightXML3/METHODNAME", [method], [data], [datatype], [callback], [errorCallback])
-
-// WebAuthentication
-// WebAuthentication
-// httpGet("http://flightxml.flightaware.com/json/FlightXML3/WeatherConditions?airport_code=KJFK&weather_date=0&howMany=1&offset=0");
-
-// let url = http://flightxml.flightaware.com/json/FlightXML3/METHODNAME
-// Username = "********"
-// API Key = "fc2a7d607c39780ddb797b27a76572a79d82ff12"
-// fxmlUrl = "https://flightxml.flightaware.com/json/FlightXML3/"
-
-// let url = 'https://flightxml.flightaware.com/json/FlightXML3/AirportBoards?airport_code=CYXE'
-
 function preload() {
   cyyz = loadJSON('assets/Toronto.json')
   cyxe = loadJSON('assets/Saskatoon.json')
@@ -53,6 +38,8 @@ function preload() {
   omdb = loadJSON('assets/Dubai.json')
   yssy = loadJSON('assets/Sydney.json')
   
+
+
   
   codes = loadStrings('assets/AirlineCodes.txt')
   font_1 = loadFont("assets/open-24-display/Open 24 Display St.ttf");
@@ -61,41 +48,47 @@ function preload() {
 }
 
 function setup() {
-  createCanvas(1700, 900);
+  createCanvas(2000, 900);
+  
   setupCityMenu();
+  info = cyxe;
   setAirlineCodes();
   cellSize = height / rows;
 
   x = 10;
   y = 45;
-  
-  
-
-  departures = info.AirportBoardsResult.departures.flights;
-  arrivals = info.AirportBoardsResult.arrivals.flights;
-  scheduled = info.AirportBoardsResult.scheduled.flights;
-  enroute = info.AirportBoardsResult.enroute.flights;
-
-  
 
 }
 
 function draw() {
   background(20);
 
+  textFont(font_1);
+  textSize(35);
+  textAlign(CENTER);
+  fill(12,56,255);
+  text("Airport Boards", width - 300, 175);
+  text ("Choose your city below", width - 300, 225);
+
+  departures = info.AirportBoardsResult.departures.flights;
+  arrivals = info.AirportBoardsResult.arrivals.flights;
+  scheduled = info.AirportBoardsResult.scheduled.flights;
+  enroute = info.AirportBoardsResult.enroute.flights;
+
   determineState();
   displayGrid();
   displayJSON();
+  noLoop();
 }
 function determineState() {
   if (state === 1) {
     status = scheduled;
-    fill(255);
+    fill(255, 137, 12);
     textFont(font_1);
     textSize(35);
     textAlign(CENTER);
-    text(info.AirportBoardsResult.airport_info.name + " Airport", 1500, height / 2 - 75);
-    text("Scheduled Flights", 1500, height/2);
+    text(info.AirportBoardsResult.airport_info.name + " Airport", width - 300, height / 2);
+    text("Scheduled Flights", width - 300, height/2 + 55);
 
     fill(200);
     textFont(font_3);
@@ -107,12 +100,12 @@ function determineState() {
   else if (state === 2) {
     status = departures;
 
-    fill(255);
+    fill(255, 137, 12);
     textFont(font_1);
     textSize(35);
     textAlign(CENTER);
-    text(info.AirportBoardsResult.airport_info.name + " Airport", 1500, height / 2 - 75);
-    text("Departed Flights", 1500, height/2);
+    text(info.AirportBoardsResult.airport_info.name + " Airport", width - 300, height / 2);
+    text("Departed Flights", width - 300, height/2 + 55);
 
     fill(200);
     textFont(font_3);
@@ -123,12 +116,12 @@ function determineState() {
   else if (state === 3) {
     status = arrivals;
 
-    fill(255);
+    fill(255, 137, 12);
     textFont(font_1);
     textSize(35);
     textAlign(CENTER);
-    text(info.AirportBoardsResult.airport_info.name + " Airport", 1500, height / 2 - 75);
-    text("Arrived Flights", 1500, height/2);
+    text(info.AirportBoardsResult.airport_info.name + " Airport", width - 300, height / 2);
+    text("Arrived Flights", width - 300, height/2 + 55);
 
     fill(200);
     textFont(font_3);
@@ -146,13 +139,20 @@ function displayJSON() {
   fill(200);
   textFont(font_3);
   textSize(28);
-  text("AIRLINE", widths[0] + 10, 40);
-  text("FLIGHT", widths[0] + widths[1]+ 10, 40);
-  text("STATUS", widths[0] + widths[1] + widths[2] + widths[3] + widths[4] + 10, 40);
-  textSize(19);
-  text("EST.TIME", widths[0] + widths[1] + widths[2] + 3, 40);
-  textSize(24);
-  text("ACTUAL", widths[0] + widths[1] + widths[2] + widths[3] + 10, 40);
+  text("AIRLINE", widths[0] + 80, 40);
+  text("FLIGHT", widths[0] + widths[1]+ 20, 40);
+  text("STATUS", widths[0] + widths[1] + widths[2] + widths[3] + widths[4] + 50, 40);
+  textSize(14);
+  text("EST.TIME", widths[0] + widths[1] + widths[2] + 10, 20);
+  text("ACTUAL TIME", widths[0] + widths[1] + widths[2] + widths[3] + 5, 22);
+  if (status === arrivals) {
+    text("ARRIVAL", widths[0] + widths[1] + widths[2] + 10, 50);
+    text("ARRIVAL", widths[0] + widths[1] + widths[2] + widths[3] + 12, 50);
+  }
+  else {
+    text("DEPARTURE", widths[0] + widths[1] + widths[2] + 10, 50);
+    text("DEPARTURE", widths[0] + widths[1] + widths[2] + widths[3] + 12, 50);
+  }
 
   
   for (let j = 0; j < cols; j++) {
@@ -193,8 +193,14 @@ function displayJSON() {
           text("Unknown Airline", widths[0] + x, y);
         }
         text(status[i].ident, widths[0] + widths[1] + x, y);
-        text(status[i].filed_departure_time.time, widths[0] + widths[1] + widths[2] + x, y);
-        text(status[i].actual_departure_time.time, widths[0] + widths[1] + widths[2] + widths[3] + x, y);
+        if (status === arrivals) {
+          text(status[i].filed_arrival_time.time, widths[0] + widths[1] + widths[2] + x, y);
+          text(status[i].actual_arrival_time.time, widths[0] + widths[1] + widths[2] + widths[3] + x, y);
+        }
+        else {
+          text(status[i].filed_departure_time.time, widths[0] + widths[1] + widths[2] + x, y);
+          text(status[i].actual_departure_time.time, widths[0] + widths[1] + widths[2] + widths[3] + x, y);
+        }
 
         text(status[i].status, widths[0] + widths[1] + widths[2] + widths[3] + widths[4] + x , y);
       }
@@ -218,7 +224,7 @@ function displayGrid() {
 }
 
 
-function mouseClicked() {
+function keyPressed() {
   if (state === 1) {
     state = 2;
   }
@@ -231,6 +237,8 @@ function mouseClicked() {
   // else if (state === 4) {
   //   state = 1;
   // }
+
+  loop();
 }
 
 function setAirlineCodes() {
@@ -246,20 +254,21 @@ function setAirlineCodes() {
 
 function setupCityMenu() {
   dropdown = createSelect();
-  dropdown.position(10, 10);
+  dropdown.position(width-350, 300);
+  dropdown.size(125,35)
   dropdown.option('Saskatoon');
   dropdown.option('Toronto');
-  // dropdown.option('Regina');
-  // dropdown.option('Calgary');
-  // dropdown.option('Vancouver');
-  // dropdown.option('Montreal');
-  // dropdown.option('Los Angeles');
-  // dropdown.option('New York (JFK)');
-  // dropdown.option('Chicago');
-  // dropdown.option('London');
-  // dropdown.option('Paris');
-  // dropdown.option('Dubai');
-  // dropdown.option('Sydney');
+  dropdown.option('Regina');
+  dropdown.option('Calgary');
+  dropdown.option('Vancouver');
+  dropdown.option('Montreal');
+  dropdown.option('Los Angeles');
+  dropdown.option('New York (JFK)');
+  dropdown.option('Chicago');
+  dropdown.option('London');
+  dropdown.option('Paris');
+  dropdown.option('Dubai');
+  dropdown.option('Sydney');
 
   dropdown.changed(mySelectEvent);
 }
@@ -271,5 +280,41 @@ function mySelectEvent() {
   else if (item === "Toronto") {
     info = cyyz;
   }
+  else if (item === "Regina") {
+    info = cyqr;
+  }
+  else if (item === "Calgary") {
+    info = cyyc;
+  }
+  else if (item === "Vancouver") {
+    info = cyvr;
+  }
+  else if (item === "Montreal") {
+    info = cyul;
+  }
+  else if (item === "Los Angeles") {
+    info = klax;
+  }
+  else if (item === "New York (JFK)") {
+    info = kjfk;
+  }
+  else if (item === "Chicago") {
+    info = kord;
+  }
+  
+  else if (item === "London") {
+    info = egll;
+  }
+  else if (item === "Paris") {
+    info = lfpg;
+  }
+  else if (item === "Dubai") {
+    info = omdb;
+  }
+  else if (item === "Sydney") {
+    info = yssy;
+  }
+  
   redraw();
+
 }