+++ history.md
@@ -3,4 +3,10 @@
 # December 11, 2018
  - Created Major Project proposal, including need to have list and nice to have list
 
+# December 20, 2018
+ - Used Regular Expression to set over 500 airline codes to the respected airline name
+
+# January 9, 2019
+ - Completed Airport Boards - Manual Edition
+